Extract LogEntryParams to dedicated file
