+++ src/schemas/index.ts
@@ -904,15 +904,8 @@ export type ToolUsage = z.infer<typeof toolUsageSchema>
  * LogEntryParams
  */
 
-export const logLevels = ["debug", "info", "warn", "error"] as const
-export const logLevelsSchema = z.enum(logLevels).optional().default("info")
-
-export const logEntryParamsSchema = z.object({
-	message: z.string(),
-	level: logLevelsSchema,
-})
-
-export type LogEntryParams = z.infer<typeof logEntryParamsSchema>
+export { logLevels, logLevelsSchema, logEntryParamsSchema } from "./log-entry"
+export type { LogEntryParams } from "./log-entry"
 
 /**
  * RooCodeEvent
